Add files via upload
Adding in a temporary bookStore file.  This contains an array of all the books I'd like the app to start with in memory.  This simulates a sort of database connection without actually making any Ajax calls or anything like that.  The idea is that the array will be altered during use of the application (such as having books added or deleted) and be passed around and filtered according to section to display the books in the correct location in the app.