Update App.js
Replaced the book array, which simulates a JSON response from an API, back into the `App.js` file for the time being.  Also added in some logic inside `ListOfBooks` to dynamically render any information stored on the book array.  The `section` attribute on each element in the array is compared against the name of the section (supplied via prop), so it can be expanded to all sections in the near future, and will be the primary factor controlling the sections displaying the correct information.