Update App.js
Transferred more of the standard HTML into react components, so now all of the sections are denoted by the `<Section>` component.  Added a `section` property to each book element.  This will need to be used to set the initial state of the book.  From there, the dropdown will handle all state changes.