Update App.js
Added in book moving functionality to app, as well as a `componentDidMount` function on the BookDropdown to be sure that the dropdown sets the selector to the correct section.  This is necessary for the `onChange` function to work correctly.  Also swapped `Book` out to be a stateless functional component and updated some functions into ES6 arrow functions.