Update App.js
Incorporated more API calls to allow for data persistence.  The application now strictly utilizes callbacks to access data that is sitting on the server rather than keeping it in local memory.  Also transitioned some additional functions to ES6 arrow functions, reworked some of the properties being passed in the hierarchy (mainly trying to reduce number of props), moved the `Section` field higher up in the code closer to its first implementation, and switched the `values` of the dropdown to reflect the JSON response from the server in order to compare against it in the `filter()` call inside of `Section`.